+++ src/main/java/net/geforcemods/securitycraft/gui/GuiProjector.java
@@ -1,12 +1,18 @@
 package net.geforcemods.securitycraft.gui;
 
+import java.awt.Rectangle;
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.containers.ContainerProjector;
+import net.geforcemods.securitycraft.containers.ContainerStateSelectorAccess;
 import net.geforcemods.securitycraft.gui.components.ClickButton;
 import net.geforcemods.securitycraft.gui.components.GuiSlider;
 import net.geforcemods.securitycraft.gui.components.GuiSlider.ISlider;
+import net.geforcemods.securitycraft.gui.components.StateSelector;
 import net.geforcemods.securitycraft.gui.components.StringHoverChecker;
 import net.geforcemods.securitycraft.gui.components.TogglePictureButton;
 import net.geforcemods.securitycraft.network.server.SyncProjector;
@@ -34,39 +40,42 @@ public class GuiProjector extends GuiContainer implements ISlider {
 	private GuiSlider projectionRangeSlider;
 	private GuiSlider projectionOffsetSlider;
 	private TogglePictureButton toggleButton;
+	private StateSelector stateSelector;
 	private int sliderWidth = 120;
 
 	public GuiProjector(InventoryPlayer inv, TileEntityProjector te) {
 		super(new ContainerProjector(inv, te));
 		this.te = te;
 		blockName = Utils.localize(te.getWorld().getBlockState(te.getPos()).getBlock().getTranslationKey() + ".name").getFormattedText();
-		ySize = 225;
+		ySize = 235;
 	}
 
 	@Override
 	public void initGui() {
 		super.initGui();
+		guiLeft += 90;
+
 		int id = 0;
 		int left = guiLeft + ((xSize - sliderWidth) / 2);
 
-		projectionWidthSlider = addButton(new GuiSlider(Utils.localize("gui.securitycraft:projector.width", te.getProjectionWidth()).getFormattedText(), blockName, id, left, guiTop + 47, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.width", "").getFormattedText(), TileEntityProjector.MIN_WIDTH, TileEntityProjector.MAX_WIDTH, te.getProjectionWidth(), false, true, this));
-		projectionWidthSlider.packedFGColour = 14737632;
+		projectionWidthSlider = addButton(new GuiSlider(Utils.localize("gui.securitycraft:projector.width", te.getProjectionWidth()).getFormattedText(), blockName, id, left, guiTop + 57, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.width", "").getFormattedText(), TileEntityProjector.MIN_WIDTH, TileEntityProjector.MAX_WIDTH, te.getProjectionWidth(), false, true, this));
+		projectionWidthSlider.packedFGColour = 0xE0E0E0;
 		hoverCheckers[id++] = new StringHoverChecker(projectionWidthSlider, Utils.localize("gui.securitycraft:projector.width.description").getFormattedText());
 
-		projectionHeightSlider = addButton(new GuiSlider(Utils.localize("gui.securitycraft:projector.height", te.getProjectionHeight()).getFormattedText(), blockName, id, left, guiTop + 68, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.height", "").getFormattedText(), TileEntityProjector.MIN_WIDTH, TileEntityProjector.MAX_WIDTH, te.getProjectionHeight(), false, true, this));
-		projectionHeightSlider.packedFGColour = 14737632;
+		projectionHeightSlider = addButton(new GuiSlider(Utils.localize("gui.securitycraft:projector.height", te.getProjectionHeight()).getFormattedText(), blockName, id, left, guiTop + 78, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.height", "").getFormattedText(), TileEntityProjector.MIN_WIDTH, TileEntityProjector.MAX_WIDTH, te.getProjectionHeight(), false, true, this));
+		projectionHeightSlider.packedFGColour = 0xE0E0E0;
 		hoverCheckers[id++] = new StringHoverChecker(projectionHeightSlider, Utils.localize("gui.securitycraft:projector.height.description").getFormattedText());
 
-		projectionRangeSlider = addButton(new GuiSlider(Utils.localize("gui.securitycraft:projector.range", te.getProjectionRange()).getFormattedText(), blockName, id, left, guiTop + 89, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.range", "").getFormattedText(), TileEntityProjector.MIN_RANGE, TileEntityProjector.MAX_RANGE, te.getProjectionRange(), false, true, this));
-		projectionWidthSlider.packedFGColour = 14737632;
+		projectionRangeSlider = addButton(new GuiSlider(Utils.localize("gui.securitycraft:projector.range", te.getProjectionRange()).getFormattedText(), blockName, id, left, guiTop + 99, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.range", "").getFormattedText(), TileEntityProjector.MIN_RANGE, TileEntityProjector.MAX_RANGE, te.getProjectionRange(), false, true, this));
+		projectionWidthSlider.packedFGColour = 0xE0E0E0;
 		hoverCheckers[id++] = new StringHoverChecker(projectionRangeSlider, Utils.localize("gui.securitycraft:projector.range.description").getFormattedText());
 
-		projectionOffsetSlider = addButton(new GuiSlider(Utils.localize("gui.securitycraft:projector.offset", te.getProjectionOffset()).getFormattedText(), blockName, id, left, guiTop + 110, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.offset", "").getFormattedText(), TileEntityProjector.MIN_OFFSET, TileEntityProjector.MAX_OFFSET, te.getProjectionOffset(), false, true, this));
-		projectionWidthSlider.packedFGColour = 14737632;
+		projectionOffsetSlider = addButton(new GuiSlider(Utils.localize("gui.securitycraft:projector.offset", te.getProjectionOffset()).getFormattedText(), blockName, id, left, guiTop + 120, sliderWidth, 20, Utils.localize("gui.securitycraft:projector.offset", "").getFormattedText(), TileEntityProjector.MIN_OFFSET, TileEntityProjector.MAX_OFFSET, te.getProjectionOffset(), false, true, this));
+		projectionWidthSlider.packedFGColour = 0xE0E0E0;
 		hoverCheckers[id++] = new StringHoverChecker(projectionOffsetSlider, Utils.localize("gui.securitycraft:projector.offset.description").getFormattedText());
 
 		//@formatter:off
-		toggleButton = addButton(new TogglePictureButton(id, left, guiTop + 26, 20, 20, TEXTURE, new int[] {176, 192}, new int[] {0, 0}, 2, 2, b -> {
+		toggleButton = addButton(new TogglePictureButton(id, left + sliderWidth - 20, guiTop + 36, 20, 20, TEXTURE, new int[] {176, 192}, new int[] {0, 0}, 2, 2, b -> {
 			//@formatter:on
 			te.setHorizontal(!te.isHorizontal());
 			projectionRangeSlider.updateSlider();
@@ -77,6 +86,9 @@ public class GuiProjector extends GuiContainer implements ISlider {
 		projectionRangeSlider.updateSlider();
 
 		slotHoverChecker = new StringHoverChecker(guiTop + 22, guiTop + 39, guiLeft + 78, guiLeft + 95, SLOT_TOOLTIP);
+
+		stateSelector = new StateSelector((ContainerStateSelectorAccess) inventorySlots, guiLeft - 190, guiTop + 7, 0, 197, 0, -2.85F, -0.45F);
+		stateSelector.initGui();
 	}
 
 	@Override
@@ -104,15 +116,59 @@ public class GuiProjector extends GuiContainer implements ISlider {
 		drawDefaultBackground();
 		GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
 		mc.getTextureManager().bindTexture(TEXTURE);
-		int startX = (width - xSize) / 2;
-		int startY = (height - ySize) / 2;
-		this.drawTexturedModalRect(startX, startY, 0, 0, xSize, ySize);
+		drawTexturedModalRect(guiLeft, guiTop, 0, 0, xSize, ySize);
+
+		if (stateSelector != null)
+			stateSelector.drawScreen(mouseX, mouseY, partialTicks);
 	}
 
 	@Override
 	protected void actionPerformed(GuiButton button) {
 		if (button instanceof ClickButton)
 			((ClickButton) button).onClick();
+	}
+
+	@Override
+	protected void mouseClickMove(int mouseX, int mouseY, int button, long timeSinceLastClick) {
+		if (stateSelector != null) {
+			stateSelector.mouseClickMove(mouseX, mouseY, button, timeSinceLastClick);
+			return;
+		}
+
+		super.mouseClickMove(mouseX, mouseY, button, timeSinceLastClick);
+	}
+
+	@Override
+	protected void mouseClicked(int mouseX, int mouseY, int button) throws IOException {
+		if (stateSelector != null)
+			stateSelector.mouseClicked(mouseX, mouseY, button);
+
+		super.mouseClicked(mouseX, mouseY, button);
+	}
+
+	@Override
+	protected void mouseReleased(int mouseX, int mouseY, int state) {
+		if (stateSelector != null)
+			stateSelector.mouseReleased(mouseX, mouseY, state);
+
+		super.mouseReleased(mouseX, mouseY, state);
+	}
+
+	public List<Rectangle> getExtraAreas() {
+		if (stateSelector != null)
+			return stateSelector.getExtraAreas();
+		else
+			return new ArrayList<>();
+	}
+
+	@Override
+	public void onGuiClosed() {
+		super.onGuiClosed();
+
+		if (stateSelector.getState() != null) {
+			te.setProjectedState(stateSelector.getState());
+			SecurityCraft.network.sendToServer(new SyncProjector(te.getPos(), stateSelector.getState()));
+		}
 	}
 
 	@Override